edit the prommision superadmin

+++ app/Http/Resources/ProductResource.php
@@ -24,8 +24,8 @@ class ProductResource extends JsonResource
             'category_id' => $this->category_id,
             'unit' => $this->unit,
 
-            'price' => (float) ($this->price ?? 0),
-            'last_price' => (float) ($this->last_price ?? 0),
+            'price' => $this->price ?? 0,
+'last_price' => $this->last_price ?? 0,
             'unit_tax' => (float) ($this->tax ?? 0),
             'description' => $this->notes,
             'available_quantity' => (int) ($this->available_quantity ?? 0),
